restructure and update practical examples
