--- Practical_examples/modelling/groundwater_model/model/2017.nam
@@ -1,11 +1,0 @@
-# Name file for MODFLOW-2005, generated by Flopy version 3.2.9.
-#xul:0; yul:0; rotation:0; proj4_str:+init=EPSG:4326; units:meters; lenuni:2; length_multiplier:1.0 ;start_datetime:2017-03-01 00:00:00
-LIST               2  2017.list
-DIS               11  2017.dis 
-BAS6              13  2017.bas 
-BCF6              15  2017.bcf 
-OC                14  2017.oc 
-PCG               27  2017.pcg 
-RCH               19  2017.rch 
-RIV               18  2017.riv 
-DATA(BINARY)      51  2017.hds REPLACE
