--- Practical_examples/modelling/groundwater_model/model/2018.pcg
@@ -1,3 +1,0 @@
-# PCG package for  MODFLOW-2005, generated by Flopy.
-50 30 1 0
-1e-05 1e-05 1.0 0 0 3 1.0
