--- Practical_examples/modelling/groundwater_model/model/average_1960-2018.nam
@@ -1,11 +1,0 @@
-# Name file for MODFLOW-2005, generated by Flopy version 3.2.9.
-#xul:0; yul:0; rotation:0; proj4_str:+init=EPSG:4326; units:meters; lenuni:2; length_multiplier:1.0 ;start_datetime:2018-03-01 00:00:00
-LIST               2  average_1960-2018.list
-DIS               11  average_1960-2018.dis 
-BAS6              13  average_1960-2018.bas 
-BCF6              15  average_1960-2018.bcf 
-OC                14  average_1960-2018.oc 
-PCG               27  average_1960-2018.pcg 
-RCH               19  average_1960-2018.rch 
-RIV               18  average_1960-2018.riv 
-DATA(BINARY)      51  average_1960-2018.hds REPLACE
